Move getInternalMethods as util
