+++ src/alt/ActionCreator.js
@@ -1,4 +1,4 @@
-import { ACTION_HANDLER, ACTION_UID} from './symbols/symbols'
+import { ACTION_HANDLER, ACTION_UID } from './symbols/symbols'
 
 export default class ActionCreator {
   constructor(alt, name, action, actions) {
